move asfreq tests
